--- client/src/components/Dashboard/Dashboard.css
@@ -1,3 +1,0 @@
-body {
-	font-size: 28px;
-}
