+++ client/src/components/Footer/Footer.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Footer = () => {
+  return (
+    <footer className="w-100 mt-auto bg-secondary p-4">
+      <div className="container">&copy;2020 by Ben, James, Kyle and Mark</div>
+    </footer>
+  );
+};
+
+export default Footer;
