--- client/src/components/Login/Login.css
@@ -1,40 +1,0 @@
-body {
-    background-color: whitesmoke;
-    font-family:Arial, Helvetica, sans-serif;
-}
-
-.login-wrapper {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-h2 {
-    font-size: 36px;
-    color:#fbb040;
-    font-weight: bold;
-    text-shadow: 2px 2px #292826;
-    padding-top: 64px;
-}
-
-p {
-    font-size: 24px;
-   color: #fbb040;
-   font-weight: bold;
-   text-shadow: 2px 2px #292826;
-   
-}
-
-button {
-    background-color: #fbb040;
-    font-weight: bold;
-    font-size: 18px;
-    padding: 7px;
-    text-shadow: .5px .5px #292826;
-    box-shadow: .5px 1px #12100e;
-}
-
-.password {
-    height: 28px;
-}
-
