+++ client/src/pages/Profile.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 
-class Dashboard extends Component {
+class Profile extends Component {
 
   constructor(props) {
     super(props)
@@ -67,4 +67,4 @@ class Dashboard extends Component {
   }
 }
 
-export default Dashboard;
+export default Profile;
